Add files via upload
Akos Boda
Driving a controllerless LCD with STM32
This project is 5 years old, deals with old monochrome controllerless LCDs.
Most of the time these LCDs are (obsolete) high quality high resolution models. The
contrast is not so nice, and the backlight is usually a strange electroluminescens
sheet or CFL lamp (but this can easily be replaced by a 12V LED strip).
Why could this still be interesting : can display lot's of characters (or graphics)
and the LCD itself can be recovered for free from old LCD monitors with 640x480
or similar resolution. Last but not least it is a challenge for STM32 fans... like me.
The STM32 can handle the speed of all signals for controlling the LCD. The Arduino program
uses fast port write so for example the clock speed is around 1 Mhz. A controller like this
has the following tasks: send frame start signal, read 4 bits from RAM, put the 4 bits
on port, move clock, repeat this 640/4 times, at end of the line send line latch signal,
repeat line cycle to do all 480 lines.
For more explanation please read the code, and see the next link for info on hardware :
http://img.hqew.com/.../Elect.../20080807001540982DVDL83.PDF
The LCD main signals are : line load, frame, data 4 bit wide, clock, M signal.
They need +5V, -22V voltage for the operation, provided by a MC34063 converter.
The whole LCD data is stored in the RAM of the STM32. In the time this was developed
I only had the STM32F103 with limited RAM, that is why only half of the LCD is used.
Today this would not be a limitation any more.
It would be interesting to develop the idea further and write the RAM contents from
serial input etc. creating a full functioning graphic terminal.